[refactor] presentation - Studio 관련 view, viewModel 분리
